Changed added LaTeX commands for clearing pages, and added additional experimental commands.
The \clearversopage command is supposed to make sure that a blank verso page
follows the content last output, and will make sure following content starts on
the succeeding recto page.

Similarly, the \clearrectopage command is supposed to clear the following recto
page, and flush consecutive content to the next verso page.

The commands have been tested with several cases and (hopefully) works. The
\clearversopage command could, for instance, be used to make sure that the
document ends with a blank back cover.

Note that we would need an additional command if we desire a special back cover
page instead.