Revert back to \lineskiplimit=0 since a skip limit < 0 is a continuing cause of spacing regressions in tables, figures, and other environments.
It seems to work to set \lineskiplimit=-\maxdimen locally for chapter heading,
to make sure that consecutive normal text will be aligned at all times.

I have noted that chapter titles that span several lines and use characters
with high ascenders (such as \'A) may overlap with other lines if typesetting
with normal size = 10pt. Maybe we later have to tweak the inter-line spacing
in the chapter headers for each font size?