Enforce caption positioning for floats.
It seems that common practice and consensus among typographers etc. is to place
captions *above* all kinds of floats (tables, listings, etc.) except for figures.
I.e. figures is the only kind of float where captions should be placed below the
"thing". Should we enforce this as default in the template class?