Fixed bug in glossaries customization that caused improper sorting in LoA.
Thanks to @kelsebo for first noticing the issue.
However, it turned out to be caused by my hack.
Have removed smallcaps-formatting from the `name' field, and applied it in a
custom glossary style instead (based on the default list style). The formatting
should not affect the other styles, or styling of glossaries that are not LoA.