Fix copyright bug.
Copyright notice was output even with 'nocopyright' option if title page was
first page. Also, deferred (non-immediate) writes and similar that
are setup before \begin{document} are part of shipout box for first page,
because \write, \openout, and \closeout operations that are performed without
\immediate are delayed until shipout time. Because the first page we shipout
is the copyright page, we lost e.g. \write1 commands from \makeglossaries when
we discarded copyright page in the case where front page was first effective
page. Now, we insert the original shipout box for copyright page after the
title page, instead of re-creating the copyright page.